Migrate jest mocks to more logical placement, improve gatsby mock
