+++ __mocks__/modules/gatsby-image.js
@@ -1,0 +1,12 @@
+import React from "react"
+
+jest.mock('gatsby-image', () => {
+  return (props) => {
+    return <img
+      src={props.fixed}
+      alt={props.alt}
+      data-testid={props['data-testid']}
+      className={props.className}
+    />;
+  }
+});
